Custom settings, thanks @RobertoHE et al

+++ src/BLEMIDI_Settings.h
@@ -6,7 +6,7 @@ BEGIN_BLEMIDI_NAMESPACE
 
 struct DefaultSettings
 {
-    static const size_t MaxBufferSize = 64;
+    static const short MaxBufferSize = 64;
 };
 
 END_BLEMIDI_NAMESPACE
